requirements.txt
PySide6>=6.5.0
mss>=9.0.0
opencv-python>=4.8.0
numpy>=1.24.0
scikit-learn>=1.3.0
Pillow>=10.0.0
pystray>=0.19.4
pyinstaller>=6.0.0
